try other way to show reminder but fail
