+++ src/main/java/seedu/mark/ui/MainWindow.java
@@ -2,9 +2,15 @@ package seedu.mark.ui;
 
 import static java.util.Objects.requireNonNull;
 
+import java.time.Duration;
+import java.time.LocalDateTime;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
 import java.util.logging.Logger;
 
+import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.MenuItem;
@@ -15,6 +21,7 @@ import javafx.scene.input.KeyCombination;
 import javafx.scene.input.KeyEvent;
 import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
+import org.controlsfx.control.Notifications;
 import seedu.mark.commons.core.GuiSettings;
 import seedu.mark.commons.core.LogsCenter;
 import seedu.mark.logic.Logic;
@@ -23,6 +30,7 @@ import seedu.mark.logic.commands.exceptions.CommandException;
 import seedu.mark.logic.commands.results.CommandResult;
 import seedu.mark.logic.parser.exceptions.ParseException;
 import seedu.mark.model.bookmark.Url;
+import seedu.mark.model.reminder.Reminder;
 
 /**
  * The Main Window. Provides the basic application layout containing
@@ -78,6 +86,11 @@ public class MainWindow extends UiPart<Stage> {
     @FXML
     private StackPane folderStructurePlaceholder;
 
+    private ScheduledExecutorService executor = Executors.newScheduledThreadPool (1);
+    Runnable r = () -> showDueReminder();
+
+    private ObservableList<Reminder> reminders;
+
     public MainWindow(Stage primaryStage, Logic logic) {
         super(FXML, primaryStage);
 
@@ -91,6 +104,8 @@ public class MainWindow extends UiPart<Stage> {
         setAccelerators();
 
         helpWindow = new HelpWindow();
+        reminders = logic.getReminderList();
+        executor.scheduleAtFixedRate( r , 0L , 5L , TimeUnit.SECONDS);
     }
 
     public Stage getPrimaryStage() {
@@ -303,6 +318,28 @@ public class MainWindow extends UiPart<Stage> {
             logger.info("Invalid command: " + commandText);
             resultDisplay.setFeedbackToUser(e.getMessage());
             throw e;
+        }
+    }
+
+    private long compareTime(LocalDateTime before, LocalDateTime after) {
+        return Duration.between(before, after).toHours();
+    }
+
+    private void showDueReminder() {
+        LocalDateTime now = LocalDateTime.now();
+        System.out.println("running show");
+
+        for (int i = 0; i < reminders.size(); i++) {
+            Reminder reminder = reminders.get(i);
+            LocalDateTime remindTime = reminder.getRemindTime();
+            System.out.println(compareTime(now, remindTime));
+            System.out.println(now.isBefore(remindTime) && compareTime(now, remindTime) < 5);
+            if (now.isBefore(remindTime) && compareTime(now, remindTime) < 5) {
+                System.out.println(true);
+                Notifications.create() .title("Task Reminder") .text(reminder.getNote().toString()) .show();
+//                ReminderWindow window = new ReminderWindow(reminder.getUrl(), reminder.getNote());
+//                window.show();
+            }
         }
     }
 }
